More Classes + Projectile Interactions
Added LivingMutablePhysicsObject and NonlivingMutablePhysicsObject (subclasses of MutablePhysicsObject), and GoodCreature and EnemyCreature (subclasses of LivingMutablePhysicsObject), made Kirby a subclass of GoodCreature, added Projectile(Subclass of NonlivingMutablePhysicsObject), GoodProjectile and BadProjectile (subclasses of Projectile), and added PhysicsObjectSet.

Projectiles have their own methods for interacting with walls and creatures. A projectile allied with a creature will currently weirdly bounce off. A projectile opposed to a creature will disappear, and deal the creature damage, although currently the damage dealt is buggily too much. (Probably collision interactions.) Collisions now checks if an object colliding is a projectile before it does does its things.

PhysicsObjectSet is a new helper class that removes the need for adding each object added to it to the same scene and the Physics enginething each time. It also provides a way for objects to effectively delete themselves, by calling getSet().remove(this);
Just create the PhysicsObjectSet using the scene and group you're using, and add all objects you're using to it. It takes care of the rest. (Except Kirby. You still need to activateControls(), although you don't need to input the scene for it to work now)

TODO: Kirby is still falling through the block at high velocities. Need to fix that. We can start creating subclasses for specific enemy and block types soon.